% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\docType{data}
\name{mid}
\alias{mid}
\title{Corrected mass isotope distributions}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1586 rows and 12 columns.
}
\usage{
mid
}
\description{
A dataset containing the raw peak areas corrected for quadrupole bias and
natural isotope abudnance.
}
\details{
\describe{
\item{mode}{
\code{fs} = full scan \cr
\code{sim} = selected ion monitoring}
\item{date}{LC-MS run date}
\item{filename}{sample run index}
\item{replicate}{biological replicate}
\item{treatment}{
\code{ctl} = control \cr
\code{lps} = lipopolysaccharide \cr
\code{tnf} = tumor necrosis factor α}
\item{metabolite}{metabolite name}
\item{isotope}{number of carbon-13 isotopes in the molecule}
\item{area}{raw peak area from TraceFinder software}
\item{cf}{correction factor for quadrupole bias in SIM mode}
\item{area_corr}{peak area corrected for quadrupole bias}
\item{mid}{mass isotope distribution of correct peak areas}
\item{mid_corr}{MID corrected for natural isotope abundance}
}
}
\keyword{datasets}
