% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Xiao.2021.endothelial.immunometabolism-package.R
\docType{package}
\name{Xiao.2021.endothelial.immunometabolism-package}
\alias{Xiao.2021.endothelial.immunometabolism}
\alias{Xiao.2021.endothelial.immunometabolism-package}
\title{Xiao.2021.endothelial.immunometabolism: 2021 Xiao Endothelial Immunometabolic Phenotypes LC-MS Data}
\description{
This is a data package containing the raw peak intensity data and 
    analytical code for processing the LC-MS data obtained from stable isotope
    tracing experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oldhamlab/Xiao.2021.endothelial.immunometabolism}
  \item Report bugs at \url{https://github.com/oldhamlab/Xiao.2021.endothelial.immunometabolism/issues}
}

}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
